pyproject.toml
[project]
name = "resume-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "docx>=0.2.4",
    "dotenv>=0.9.9",
    "google-genai>=1.0.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.1.0",
]
